Remove LoadingIcon (Mantine has LoadingOverlay and Loader) also remove StatRenderBundle
